service/pyproject.toml
[tool.poetry]
name = "fexservice"
version = "0.1.0"
description = ""
authors = ["Bruno Rocha <rochacbruno@gmail.com>"]

[tool.poetry.scripts]
fexservice = "fexservice.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
dynaconf = "^2.2.3"
pygithub = "^1.47"
sqlalchemy = "^1.3.15"
dataset = "^1.3.0"
typer = {version = "^0.1.0", extras = ["all"]}

[tool.poetry.dev-dependencies]
pytest = "^4.6"
ipython = "^7.13.0"
ipdb = "^0.13.2"
isort = "^4.3.21"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
